Point origin and basic Kinematic
origin property for adding relative root for its position.
if a and b are Point and b is origin of a, then a's absolute will be sum
of their positions.
b's absolute is just its position, as its origin is null.
Basic Kinematic does just adding velocity to position.
extending Visual class with Kinematic, we can see visualy.